libros buscados se registran junto con sus autores en la BD

+++ literalura/src/main/java/com/martin/literalura/model/DatosLibro.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record DatosLibro(
-        @JsonAlias("id") Integer id,
+        @JsonAlias("id") Long id,
         @JsonAlias("title") String titulo,
         @JsonAlias("authors") List<DatosAutor> autores,
         @JsonAlias("summaries") List<String> resumenes,
